refactor(engine): S2 Stage 2 — extract MIDI Import domain to XlethSvcMidi TU
Verbatim move of 3 midi_* handlers into a dedicated translation unit.
Promoted syncMixerTrackSlots + triggerMipmapGeneration to external
linkage (new XlethSvcHelpers.h) so XlethSvcMidi.cpp can call them.
Zero behavior change; ctest 43/47 identical to Stage-1 baseline.
